refactor(chat-composer): streamline chat composer components and enhance attachment handling
- Refactored ChatComposer to utilize a custom hook for managing attachments, improving code organization and readability.
- Introduced new components: ComposerAttachments, ComposerTextarea, and ComposerToolbar for better separation of concerns.
- Simplified file attachment logic and error handling, enhancing user experience during file uploads.
- Updated utility functions for file size formatting and attachment validation, ensuring consistent behavior across components.
